Adding finance calculator script

+++ python_introduction/finance_calculator.py
@@ -1,10 +1,9 @@
 #A finance calculator to enable user to calculate their savings over a one month period
 monthly_income= input("Enter your monthly income: ")
 
-
 monthly_expenses= input("Enter your monhtly expenses: ")
 
-monthly_savings = (float(monthly_income) - float(monthly_expenses))
+monthly_savings = (int(monthly_income) - int(monthly_expenses))
 
 projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
